Moved navbar routes into new file
